Move new distance functions to geocalc.py #370
- move functions from ichnaea.util to ichanea.geocalc
- move tests
- update docs
