+++ docs/internal_api/geocalc.rst
@@ -6,11 +6,21 @@
 Functions
 +++++++++
 
+.. autofunction:: add_meters_to_latitude
+
+.. autofunction:: add_meters_to_longitude
+
+.. autofunction:: bound
+
 .. autofunction:: centroid
+
+.. autofunction:: country_subunits_by_iso_code
 
 .. autofunction:: distance
 
 .. autofunction:: location_is_in_country
+
+.. autofunction:: math
 
 .. autofunction:: maximum_country_radius
